+++ src/app/(public)/api/auth/callback/kakao/route.ts
@@ -1,21 +1,5 @@
 import { NextRequest, NextResponse } from 'next/server'
 
-const LOGIN_PATH = '/login'
-const KAKAO_COMPLETE_PATH = '/auth/callback/kakao/complete'
-const REDIRECT_BASE_URL = process.env.NEXT_PUBLIC_SERVER_URL ?? process.env.NEXT_PUBLIC_SITE_URL
-
-const createRedirectUrl = (path: string, request: NextRequest) => {
-  if (REDIRECT_BASE_URL) {
-    try {
-      return new URL(path, REDIRECT_BASE_URL)
-    } catch {
-      // Fall through to request origin.
-    }
-  }
-
-  return new URL(path, request.nextUrl.origin)
-}
-
 export async function GET(req: NextRequest) {
   const url = req.nextUrl
   const code = url.searchParams.get('code')
@@ -26,7 +10,7 @@ export async function GET(req: NextRequest) {
   const errorDescription = url.searchParams.get('error_description')
 
   if (error) {
-    const redirectUrl = createRedirectUrl(LOGIN_PATH, req)
+    const redirectUrl = new URL('/login', req.url)
     redirectUrl.searchParams.set('error', error)
     if (errorDescription) redirectUrl.searchParams.set('error_description', errorDescription)
 
@@ -36,13 +20,13 @@ export async function GET(req: NextRequest) {
   }
 
   if (!code) {
-    const res = NextResponse.redirect(createRedirectUrl(LOGIN_PATH, req))
+    const res = NextResponse.redirect(new URL('/login', req.url))
     res.cookies.set('kakao_oauth_state', '', { maxAge: 0, path: '/' })
     return res
   }
 
   if (!returnedState || !storedState || returnedState !== storedState) {
-    const redirectUrl = createRedirectUrl(LOGIN_PATH, req)
+    const redirectUrl = new URL('/login', req.url)
     redirectUrl.searchParams.set('error', 'invalid_state')
 
     const res = NextResponse.redirect(redirectUrl)
@@ -50,7 +34,7 @@ export async function GET(req: NextRequest) {
     return res
   }
 
-  const redirectUrl = createRedirectUrl(KAKAO_COMPLETE_PATH, req)
+  const redirectUrl = new URL('/auth/callback/kakao/complete', req.url)
   redirectUrl.searchParams.set('code', code)
 
   const res = NextResponse.redirect(redirectUrl)
