+++ src/features/auth/index.ts
@@ -1,3 +1,3 @@
 export { LoginButton } from './ui/client/LoginButton'
 export { useAuthStore } from './model/client/useAuthStore'
-export { useAccessToken, useClearAccesstoken, useSetAccessToken } from './model/client/useAuthStore'
+export { useAccessToken } from './model/client/useAuthStore'
