+++ src/features/provider/ui/AuthBootStrap.tsx
@@ -3,7 +3,11 @@
 import { useRouter } from 'next/navigation'
 import { useEffect } from 'react'
 
-import { useClearAccesstoken, useSetAccessToken, useAccessToken } from '@/features/auth'
+import {
+  useClearAccesstoken,
+  useSetAccessToken,
+  useAccessToken,
+} from '@/features/auth/model/client/useAuthStore'
 
 const REFRESH_PATH = '/api/auth/refresh'
 const SERVER_URL = process.env.NEXT_PUBLIC_SERVER_URL ?? ''
