+++ src/shared/index.ts
@@ -23,19 +23,6 @@ export {
   DialogClose,
 } from './ui/dialog'
 export { Tabs, TabsContent, TabsList, TabsTrigger } from './ui/tabs'
+export { LevelUpHeader } from './ui/LevelUpHeader'
 export { GoogleAnalytics } from './lib/GoogleAnalytics'
 export { Button } from './ui/button'
-export { StatusMessage } from './ui/StatusMessage'
-export { default as defaultAvatar } from './assets/images/default-avatar.svg'
-export { createUuidForRegex } from './lib/createUuidForRegex'
-export {
-  Drawer,
-  DrawerContent,
-  DrawerDescription,
-  DrawerFooter,
-  DrawerHeader,
-  DrawerTitle,
-} from './ui/drawer'
-export { Spinner } from './ui/Spinner'
-export { Field, FieldLabel, FieldDescription } from './ui/field'
-export { Input } from './ui/input'
