--- src/shared/lib/createUuidForRegex.ts
@@ -1,18 +1,0 @@
-export const createUuidForRegex = (): string => {
-  if (typeof crypto?.randomUUID === 'function') {
-    return crypto.randomUUID().toLowerCase()
-  }
-
-  if (typeof crypto?.getRandomValues === 'function') {
-    const bytes = new Uint8Array(16)
-    crypto.getRandomValues(bytes)
-
-    bytes[6] = (bytes[6] & 0x0f) | 0x40
-    bytes[8] = (bytes[8] & 0x3f) | 0x80
-
-    const hex = Array.from(bytes, (b) => b.toString(16).padStart(2, '0')).join('')
-    return `${hex.slice(0, 8)}-${hex.slice(8, 12)}-${hex.slice(12, 16)}-${hex.slice(16, 20)}-${hex.slice(20)}`
-  }
-
-  return ''
-}
